garmcloud-functions-python/requirements.txt
azure-functions
fitdecode
requests